refactor: grouped filter with settings internally
