+++ lib/src/settings/filter_model.dart
